+++ website_menu_by_user_status/tests/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 
-from . import test_logo
+from . import test_website_menu
